Created generate.rb
Created the generate script to go along with the train.rb script. This
one uses the files generated using the train.rb script to create
sentences based on a start word. Also changed the output filenames to
use either a .yaml or .json extention for easier differentiation.
Deleted unneeded files